+++ packages/core/src/prompts/multi-select.ts
@@ -1,13 +1,19 @@
+import { findCursor } from '../utils/cursor.js';
 import Prompt, { type PromptOptions } from './prompt.js';
 
-interface MultiSelectOptions<T extends { value: any }>
+interface OptionLike {
+	value: any;
+	disabled?: boolean;
+}
+
+interface MultiSelectOptions<T extends OptionLike>
 	extends PromptOptions<T['value'][], MultiSelectPrompt<T>> {
 	options: T[];
 	initialValues?: T['value'][];
 	required?: boolean;
 	cursorAt?: T['value'];
 }
-export default class MultiSelectPrompt<T extends { value: any }> extends Prompt<T['value'][]> {
+export default class MultiSelectPrompt<T extends OptionLike> extends Prompt<T['value'][]> {
 	options: T[];
 	cursor = 0;
 
@@ -15,9 +21,14 @@ export default class MultiSelectPrompt<T extends { value: any }> extends Prompt<
 		return this.options[this.cursor].value;
 	}
 
+	private get _enabledOptions(): T[] {
+		return this.options.filter((option) => option.disabled !== true);
+	}
+
 	private toggleAll() {
-		const allSelected = this.value !== undefined && this.value.length === this.options.length;
-		this.value = allSelected ? [] : this.options.map((v) => v.value);
+		const enabledOptions = this._enabledOptions;
+		const allSelected = this.value !== undefined && this.value.length === enabledOptions.length;
+		this.value = allSelected ? [] : enabledOptions.map((v) => v.value);
 	}
 
 	private toggleInvert() {
@@ -25,7 +36,7 @@ export default class MultiSelectPrompt<T extends { value: any }> extends Prompt<
 		if (!value) {
 			return;
 		}
-		const notSelected = this.options.filter((v) => !value.includes(v.value));
+		const notSelected = this._enabledOptions.filter((v) => !value.includes(v.value));
 		this.value = notSelected.map((v) => v.value);
 	}
 
@@ -44,10 +55,11 @@ export default class MultiSelectPrompt<T extends { value: any }> extends Prompt<
 
 		this.options = opts.options;
 		this.value = [...(opts.initialValues ?? [])];
-		this.cursor = Math.max(
+		const cursor = Math.max(
 			this.options.findIndex(({ value }) => value === opts.cursorAt),
 			0
 		);
+		this.cursor = this.options[cursor].disabled ? findCursor<T>(cursor, 1, this.options) : cursor;
 		this.on('key', (char) => {
 			if (char === 'a') {
 				this.toggleAll();
@@ -61,11 +73,11 @@ export default class MultiSelectPrompt<T extends { value: any }> extends Prompt<
 			switch (key) {
 				case 'left':
 				case 'up':
-					this.cursor = this.cursor === 0 ? this.options.length - 1 : this.cursor - 1;
+					this.cursor = findCursor<T>(this.cursor, -1, this.options);
 					break;
 				case 'down':
 				case 'right':
-					this.cursor = this.cursor === this.options.length - 1 ? 0 : this.cursor + 1;
+					this.cursor = findCursor<T>(this.cursor, 1, this.options);
 					break;
 				case 'space':
 					this.toggleValue();